# internal/handler/docs/swagger.yaml
basePath: /
definitions:
  calendar.Event:
    properties:
      date:
        type: string
      event:
        type: string
      id:
        type: integer
      user_id:
        type: integer
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для календаря
  title: Calendar API
  version: "1.0"
paths:
  /events:
    get:
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: integer
      - description: Date
        in: query
        name: date
        type: string
      - description: day|week|month
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/calendar.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить события
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Создает новое событие календаря
      parameters:
      - description: Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/calendar.Event'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/calendar.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Создать событие
      tags:
      - events
  /events/{id}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Удалить событие
      tags:
      - events
    put:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/calendar.Event'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/calendar.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Обновить событие
      tags:
      - events
swagger: "2.0"